+ fixed bug with wrong titlebar after typing notify in mixed containers (group chats + normal sessions) + moved some options. channel logging is now on the group chat option page and no longer with the fonts
git-svn-id: http://miranda.googlecode.com/svn/trunk@5112 83df23c2-51fd-11de-bb24-2d5e9b4af41d

+++ miranda/plugins/tabsrmm/msgdlgutils.c
@@ -1059,7 +1059,7 @@ void FlashOnClist(HWND hwndDlg, struct MessageWindowData *dat, HANDLE hEvent, DB
 
     dat->dwTickLastEvent = GetTickCount();
     if((GetForegroundWindow() != dat->pContainer->hwnd || dat->pContainer->hwndActive != hwndDlg) && !(dbei->flags & DBEF_SENT) && dbei->eventType == EVENTTYPE_MESSAGE) {
-        UpdateTrayMenu(dat, dat->bIsMeta ? dat->wMetaStatus : dat->wStatus, dat->bIsMeta ? dat->szMetaProto : dat->szProto, dat->szStatus, dat->hContact, 0L);
+        UpdateTrayMenu(dat, (WORD)(dat->bIsMeta ? dat->wMetaStatus : dat->wStatus), dat->bIsMeta ? dat->szMetaProto : dat->szProto, dat->szStatus, dat->hContact, 0L);
         if(nen_options.bTraySupport == TRUE && myGlobals.m_WinVerMajor >= 5)
             return;
     }
